refactor(pty): split session actions into interact_terminal_session tool
Separates PTY session interaction from command execution for clearer
model mental model:

- run_terminal_cmd: now only handles exec (interactive or non-interactive)
- interact_terminal_session: new tool for send/wait/view/kill on existing
  sessions (requires session ID from interactive exec)

Changes:
- Extract shared wait utilities to lib/ai/tools/utils/pty-wait-utils.ts
- Create lib/ai/tools/interact-terminal-session.ts with session actions
- Simplify run-terminal-cmd.ts schema (remove action enum, session, input)
- Add UI support in MessagePartHandler, SharedMessagePartHandler,
  message-processor, and sidebar-utils
- Move PTY tests to interact-terminal-session.test.ts
