+++ lib/ai/tools/__tests__/run-terminal-cmd.test.ts
@@ -1,17 +1,15 @@
 /**
- * Tests for `run_terminal_cmd` — focusing on the interactive PTY action
- * dispatch added by the PTY MVP.
+ * Tests for `run_terminal_cmd` — focusing on exec and interactive session creation.
  *
  * The non-interactive (`action=exec`, `interactive=false`) path is already
  * covered by higher-level integration tests. Here we verify:
- *  - the dispatch contract for {exec+interactive, send, wait, view, kill}
- *  - input size cap
- *  - pty-keys decoding (Enter / C-c / plain text)
- *  - wait policy (pattern + idle)
- *  - ANSI stripping + bufferTruncated surfacing
+ *  - the dispatch contract for {exec, exec+interactive}
  *  - structured errors for non-E2B sandboxes and missing sessions
  *  - that the legacy schema ({command, explanation, is_background, timeout})
  *    still flows through and produces a shaped result.
+ *
+ * PTY session action tests (send, wait, view, kill) are in
+ * interact-terminal-session.test.ts.
  */
 
 // Stub out @e2b/code-interpreter — its ESM `chalk` dependency trips Jest's
@@ -381,391 +379,6 @@ describe("run_terminal_cmd — PTY action dispatch", () => {
     ).toBeDefined();
   });
 
-  test("send on unknown session returns structured error", async () => {
-    const { context } = makeContext({ sandbox: makeFakeE2BSandbox() });
-    const tool = createRunTerminalCmd(context);
-    const result = (await runTool(tool, {
-      action: "send",
-      session: "nope",
-      input: "hi\n",
-      explanation: "x",
-      is_background: false,
-    })) as { result: { error?: string } };
-    expect(result.result.error).toMatch(/Session nope not found/);
-  });
-
-  test("send with oversized input errors without calling sendInput", async () => {
-    const e2b = makeFakeE2BSandbox();
-    const handle = makeFakeHandle();
-    mockCreateE2BPtyHandle.mockImplementation(async () => handle);
-
-    const { context } = makeContext({ sandbox: e2b });
-    const tool = createRunTerminalCmd(context);
-
-    // First create a session.
-    const execPromise = runTool(tool, {
-      action: "exec",
-      command: "sh",
-      explanation: "x",
-      is_background: false,
-      interactive: true,
-      wait_for: { idle_ms: 10, timeout_ms: 500 },
-    });
-    await new Promise((r) => setTimeout(r, 0));
-    const created = (await execPromise) as {
-      result: { session: string };
-    };
-    const sessionId = created.result.session;
-
-    // Count of sendInput calls so far (1 — the initial command).
-    const before = handle.sendInputCalls.length;
-
-    const huge = "a".repeat(8 * 1024 + 1);
-    const result = (await runTool(tool, {
-      action: "send",
-      session: sessionId,
-      input: huge,
-      explanation: "x",
-      is_background: false,
-    })) as { result: { error?: string } };
-
-    expect(result.result.error).toMatch(/exceeds MAX_INPUT_BYTES_PER_SEND/);
-    expect(handle.sendInputCalls.length).toBe(before);
-  });
-
-  test("send decodes tmux-style keys: C-c -> \\x03, Enter -> \\r, plain text verbatim", async () => {
-    const e2b = makeFakeE2BSandbox();
-    const handle = makeFakeHandle();
-    mockCreateE2BPtyHandle.mockImplementation(async () => handle);
-
-    const { context } = makeContext({ sandbox: e2b });
-    const tool = createRunTerminalCmd(context);
-
-    const created = (await (async () => {
-      const p = runTool(tool, {
-        action: "exec",
-        command: "sh",
-        explanation: "x",
-        is_background: false,
-        interactive: true,
-        wait_for: { idle_ms: 10, timeout_ms: 500 },
-      });
-      await new Promise((r) => setTimeout(r, 0));
-      return p;
-    })()) as { result: { session: string } };
-    const sessionId = created.result.session;
-
-    const sendAndGet = async (input: string) => {
-      const beforeLen = handle.sendInputCalls.length;
-      void runTool(tool, {
-        action: "send",
-        session: sessionId,
-        input,
-        explanation: "x",
-        is_background: false,
-        wait_for: { idle_ms: 10, timeout_ms: 200 },
-      });
-      // Yield so the tool's awaited sendInput runs.
-      await new Promise((r) => setTimeout(r, 0));
-      await new Promise((r) => setTimeout(r, 0));
-      return handle.sendInputCalls[beforeLen];
-    };
-
-    const ctrlC = await sendAndGet("C-c");
-    expect(Array.from(ctrlC)).toEqual([0x03]);
-
-    const enter = await sendAndGet("Enter");
-    expect(Array.from(enter)).toEqual([0x0d]);
-
-    const plain = await sendAndGet("hello");
-    expect(new TextDecoder().decode(plain)).toBe("hello");
-  });
-
-  test("wait resolves early when wait_for.pattern matches", async () => {
-    const e2b = makeFakeE2BSandbox();
-    const handle = makeFakeHandle();
-    mockCreateE2BPtyHandle.mockImplementation(async () => handle);
-
-    const { context } = makeContext({ sandbox: e2b });
-    const tool = createRunTerminalCmd(context);
-
-    const execP = runTool(tool, {
-      action: "exec",
-      command: "read x",
-      explanation: "x",
-      is_background: false,
-      interactive: true,
-      wait_for: { idle_ms: 5, timeout_ms: 200 },
-    });
-    await new Promise((r) => setTimeout(r, 0));
-    const created = (await execP) as { result: { session: string } };
-
-    const startedAt = Date.now();
-    const waitP = runTool(tool, {
-      action: "wait",
-      session: created.result.session,
-      explanation: "x",
-      is_background: false,
-      wait_for: { pattern: "READY", timeout_ms: 2000, idle_ms: 5 },
-    });
-    setTimeout(() => handle.emit(new TextEncoder().encode("hello READY\n")), 5);
-    const result = (await waitP) as { result: { output: string } };
-    expect(result.result.output).toContain("READY");
-    expect(Date.now() - startedAt).toBeLessThan(500);
-  });
-
-  test("wait resolves on idle after idle_ms with no new output", async () => {
-    const e2b = makeFakeE2BSandbox();
-    const handle = makeFakeHandle();
-    mockCreateE2BPtyHandle.mockImplementation(async () => handle);
-
-    const { context } = makeContext({ sandbox: e2b });
-    const tool = createRunTerminalCmd(context);
-
-    const execP = runTool(tool, {
-      action: "exec",
-      command: "sh",
-      explanation: "x",
-      is_background: false,
-      interactive: true,
-      wait_for: { idle_ms: 5, timeout_ms: 200 },
-    });
-    await new Promise((r) => setTimeout(r, 0));
-    const created = (await execP) as { result: { session: string } };
-
-    const started = Date.now();
-    const result = (await runTool(tool, {
-      action: "wait",
-      session: created.result.session,
-      explanation: "x",
-      is_background: false,
-      wait_for: { idle_ms: 30, timeout_ms: 1000 },
-    })) as { result: { output: string } };
-    expect(Date.now() - started).toBeGreaterThanOrEqual(25);
-    expect(Date.now() - started).toBeLessThan(500);
-    expect(typeof result.result.output).toBe("string");
-  });
-
-  test("view returns snapshot without advancing readCursor", async () => {
-    const e2b = makeFakeE2BSandbox();
-    const handle = makeFakeHandle();
-    mockCreateE2BPtyHandle.mockImplementation(async () => handle);
-
-    const { context } = makeContext({ sandbox: e2b });
-    const tool = createRunTerminalCmd(context);
-
-    const execP = runTool(tool, {
-      action: "exec",
-      command: "sh",
-      explanation: "x",
-      is_background: false,
-      interactive: true,
-      wait_for: { idle_ms: 5, timeout_ms: 200 },
-    });
-    await new Promise((r) => setTimeout(r, 0));
-    const created = (await execP) as { result: { session: string } };
-    const sessionId = created.result.session;
-
-    handle.emit(new TextEncoder().encode("hello\n"));
-    // wait to drain that into the buffer
-    await new Promise((r) => setTimeout(r, 10));
-
-    const view1 = (await runTool(tool, {
-      action: "view",
-      session: sessionId,
-      explanation: "x",
-      is_background: false,
-    })) as { result: { output: string } };
-
-    // subsequent wait should still see same bytes (view did not consume them)
-    const waitRes = (await runTool(tool, {
-      action: "wait",
-      session: sessionId,
-      explanation: "x",
-      is_background: false,
-      wait_for: { idle_ms: 5, timeout_ms: 200 },
-    })) as { result: { output: string } };
-
-    expect(view1.result.output).toContain("hello");
-    expect(waitRes.result.output).toContain("hello");
-  });
-
-  test("kill closes the session and returns exitCode", async () => {
-    const e2b = makeFakeE2BSandbox();
-    const handle = makeFakeHandle();
-    mockCreateE2BPtyHandle.mockImplementation(async () => handle);
-
-    const { context, ptySessionManager } = makeContext({ sandbox: e2b });
-    const tool = createRunTerminalCmd(context);
-
-    const execP = runTool(tool, {
-      action: "exec",
-      command: "sh",
-      explanation: "x",
-      is_background: false,
-      interactive: true,
-      wait_for: { idle_ms: 5, timeout_ms: 200 },
-    });
-    await new Promise((r) => setTimeout(r, 0));
-    const created = (await execP) as { result: { session: string } };
-
-    const result = (await runTool(tool, {
-      action: "kill",
-      session: created.result.session,
-      explanation: "x",
-      is_background: false,
-    })) as { result: { exitCode: number | null } };
-
-    expect(result.result.exitCode).toBe(0);
-    expect(
-      ptySessionManager.get("chat-1", created.result.session),
-    ).toBeUndefined();
-  });
-
-  test("bufferTruncated surfaces in response when ring drops data", async () => {
-    const e2b = makeFakeE2BSandbox();
-    const handle = makeFakeHandle();
-    mockCreateE2BPtyHandle.mockImplementation(async () => handle);
-
-    const { context, ptySessionManager } = makeContext({ sandbox: e2b });
-    const tool = createRunTerminalCmd(context);
-
-    const execP = runTool(tool, {
-      action: "exec",
-      command: "cat bigfile",
-      explanation: "x",
-      is_background: false,
-      interactive: true,
-      wait_for: { idle_ms: 5, timeout_ms: 200 },
-    });
-    await new Promise((r) => setTimeout(r, 0));
-    const created = (await execP) as { result: { session: string } };
-    const sessionId = created.result.session;
-
-    // Manually flip bufferTruncated on the session (simulating ring drop).
-    const session = ptySessionManager.get("chat-1", sessionId)!;
-    (session as unknown as { bufferTruncated: boolean }).bufferTruncated = true;
-
-    const view = (await runTool(tool, {
-      action: "view",
-      session: sessionId,
-      explanation: "x",
-      is_background: false,
-    })) as { result: { bufferTruncated?: boolean } };
-    expect(view.result.bufferTruncated).toBe(true);
-  });
-
-  test("ANSI escape sequences are stripped from model-facing output", async () => {
-    const e2b = makeFakeE2BSandbox();
-    const handle = makeFakeHandle();
-    mockCreateE2BPtyHandle.mockImplementation(async () => handle);
-
-    const { context } = makeContext({ sandbox: e2b });
-    const tool = createRunTerminalCmd(context);
-
-    const execP = runTool(tool, {
-      action: "exec",
-      command: "ls --color=always",
-      explanation: "x",
-      is_background: false,
-      interactive: true,
-      wait_for: { idle_ms: 20, timeout_ms: 500 },
-    });
-    await new Promise((r) => setTimeout(r, 0));
-    // ANSI red + "HI" + reset
-    handle.emit(new TextEncoder().encode("\x1b[31mHI\x1b[0m\n"));
-    const result = (await execP) as { result: { output: string } };
-    expect(result.result.output).toContain("HI");
-    expect(result.result.output).not.toContain("\x1b[31m");
-    expect(result.result.output).not.toContain("\x1b[0m");
-  });
-
-  // ── FIX 1 — invalid wait_for.pattern returns structured error ───────
-  test("wait with invalid regex returns structured error, does not crash", async () => {
-    const e2b = makeFakeE2BSandbox();
-    const handle = makeFakeHandle();
-    mockCreateE2BPtyHandle.mockImplementation(async () => handle);
-
-    const { context } = makeContext({ sandbox: e2b });
-    const tool = createRunTerminalCmd(context);
-
-    // Open a session
-    const execP = runTool(tool, {
-      action: "exec",
-      command: "sh",
-      explanation: "x",
-      is_background: false,
-      interactive: true,
-      wait_for: { idle_ms: 5, timeout_ms: 200 },
-    });
-    await new Promise((r) => setTimeout(r, 0));
-    const created = (await execP) as { result: { session: string } };
-
-    const result = (await runTool(tool, {
-      action: "wait",
-      session: created.result.session,
-      explanation: "x",
-      is_background: false,
-      wait_for: { pattern: "[unclosed", idle_ms: 100, timeout_ms: 200 },
-    })) as { result: { error?: string } };
-
-    expect(result.result.error).toMatch(/Invalid wait_for\.pattern/);
-  });
-
-  // ── FIX 2 — zod caps on wait_for durations ──────────────────────────
-  test("wait_for.idle_ms above 60000 cap is rejected by zod", async () => {
-    const e2b = makeFakeE2BSandbox();
-    const { context } = makeContext({ sandbox: e2b });
-    const tool = createRunTerminalCmd(context);
-
-    // zod validation happens inside the ai SDK's tool wrapper — calling
-    // execute() directly here bypasses zod, so instead we parse the schema
-    // manually via the tool's inputSchema.
-    const schema = (
-      tool as unknown as {
-        inputSchema: {
-          safeParse: (v: unknown) => { success: boolean; error?: unknown };
-        };
-      }
-    ).inputSchema;
-    const resultOver = schema.safeParse({
-      action: "wait",
-      session: "s",
-      explanation: "x",
-      is_background: false,
-      wait_for: { idle_ms: 120_000, timeout_ms: 200 },
-    });
-    expect(resultOver.success).toBe(false);
-
-    const resultOk = schema.safeParse({
-      action: "wait",
-      session: "s",
-      explanation: "x",
-      is_background: false,
-      wait_for: { idle_ms: 10_000, timeout_ms: 200 },
-    });
-    expect(resultOk.success).toBe(true);
-  });
-
-  test("wait_for.timeout_ms above 300000 cap is rejected by zod", async () => {
-    const e2b = makeFakeE2BSandbox();
-    const { context } = makeContext({ sandbox: e2b });
-    const tool = createRunTerminalCmd(context);
-    const schema = (
-      tool as unknown as {
-        inputSchema: { safeParse: (v: unknown) => { success: boolean } };
-      }
-    ).inputSchema;
-    const over = schema.safeParse({
-      action: "wait",
-      session: "s",
-      explanation: "x",
-      is_background: false,
-      wait_for: { idle_ms: 100, timeout_ms: 600_000 },
-    });
-    expect(over.success).toBe(false);
-  });
-
   // ── FIX 4 — factory is not invoked when cap is already hit ───────────
   test("ptySessionManager.create does NOT invoke factory when concurrency cap is hit", async () => {
     const e2b = makeFakeE2BSandbox();
@@ -820,104 +433,5 @@ describe("run_terminal_cmd — PTY action dispatch", () => {
 
     expect(result.result.error).toMatch(/spawn failed/);
     expect(ptySessionManager.list("chat-1")).toEqual([]);
-  });
-
-  // ── action=wait when process already exited ──────────────────────────
-  test("action=wait returns exited with exitCode when process already exited", async () => {
-    const e2b = makeFakeE2BSandbox();
-    // Build a handle whose exited promise is *already* settled at creation
-    // time. We do this by creating the handle, resolving its exit immediately,
-    // then shimming the PtySessionManager so its `.then` cleanup doesn't fire
-    // before we call `action=wait`.
-    const handle = makeFakeHandle();
-    mockCreateE2BPtyHandle.mockImplementation(async () => handle);
-
-    const { context, ptySessionManager } = makeContext({ sandbox: e2b });
-    const tool = createRunTerminalCmd(context);
-
-    // Create a session
-    const execP = runTool(tool, {
-      action: "exec",
-      command: "sh",
-      explanation: "x",
-      is_background: false,
-      interactive: true,
-      wait_for: { idle_ms: 5, timeout_ms: 200 },
-    });
-    await new Promise((r) => setTimeout(r, 0));
-    const created = (await execP) as { result: { session: string } };
-    const sessionId = created.result.session;
-
-    // Resolve the exit — the handle.exited promise settles.
-    handle.resolveExit(42);
-
-    // Immediately start the wait call BEFORE yielding to microtasks — the
-    // session is still in the manager because removeSession runs in a
-    // .then() microtask that hasn't fired yet.
-    const waitP = runTool(tool, {
-      action: "wait",
-      session: sessionId,
-      explanation: "x",
-      is_background: false,
-      wait_for: { idle_ms: 10, timeout_ms: 500 },
-    }) as Promise<{
-      result: {
-        exited?: { exitCode: number | null };
-        output?: string;
-        error?: string;
-      };
-    }>;
-
-    const result = await waitP;
-
-    // If the session was still reachable, result.exited should be surfaced.
-    // If the session was already removed, we get a "Session not found" error.
-    // Either path is acceptable — the critical thing is we don't hang or crash.
-    if (result.result.error) {
-      // Session was cleaned up before wait ran — verify it's the expected error
-      expect(result.result.error).toMatch(/not found/i);
-    } else {
-      // The wait captured the already-exited state
-      expect(result.result.exited).toBeDefined();
-      expect(result.result.exited!.exitCode).toBe(42);
-    }
-  });
-
-  // ── FIX 6 — sendInput rejection surfaces as structured error ─────────
-  test("send returns structured error when handle.sendInput rejects", async () => {
-    const e2b = makeFakeE2BSandbox();
-    const handle = makeFakeHandle();
-    mockCreateE2BPtyHandle.mockImplementation(async () => handle);
-
-    const { context } = makeContext({ sandbox: e2b });
-    const tool = createRunTerminalCmd(context);
-
-    const execP = runTool(tool, {
-      action: "exec",
-      command: "sh",
-      explanation: "x",
-      is_background: false,
-      interactive: true,
-      wait_for: { idle_ms: 5, timeout_ms: 200 },
-    });
-    await new Promise((r) => setTimeout(r, 0));
-    const created = (await execP) as { result: { session: string } };
-
-    // Now rewire sendInput to reject on the NEXT call.
-    (handle.sendInput as unknown as jest.Mock).mockImplementationOnce(
-      async () => {
-        throw new Error("pipe broken");
-      },
-    );
-
-    const result = (await runTool(tool, {
-      action: "send",
-      session: created.result.session,
-      input: "hi\n",
-      explanation: "x",
-      is_background: false,
-    })) as { result: { error?: string } };
-
-    expect(result.result.error).toMatch(/Failed to send input: pipe broken/);
   });
 });
